Update app icons and image assets
This commit updates the app icons for all resolutions and platforms, as well as adds a new image asset. The Android app icons were updated in the mipmap folders, while the iOS app icons were updated in the AppIcon.appiconset folder. The new image asset, OIP.jpg, was added to the assets/images folder. The dev dependency for flutter_icons was also updated to use the new image asset. Additionally, the SharedPreferencesPlugin was removed from the GeneratedPluginRegistrant.swift file in the macOS folder.