Add HomeScreen and OnBoardingController
This commit adds a new HomeScreen and OnBoardingController to the Music Player App. The OnBoardingController has a static method that navigates to the HomeScreen when called. The RoutesManagers file has also been updated to include the new HomeScreen route. The OnBoardingScreen has been updated to call the OnBoardingController's navigatorToHomeScreen method when the "Get Started" button is pressed.