Add new images and values, update UI in play music screen
Added new images for playlist and download buttons, and updated the assets managers file to include them. Also added new height and padding values, and updated the respective files. Updated the UI in the play music screen to use the new images and padding values. Replaced the heart icon with a queue music icon.