Add stop audio functionality to the music player controller and button
The commit adds a new method to the PlayMusicController class that allows stopping the audio playback. It also modifies the CustomButtonControllerPlayMusicScreen widget to call this method when the pause button is pressed. This improves the user experience by providing a way to stop the audio playback in addition to pausing it.