Refactor PlayMusicController and add loop functionality
This commit refactors the PlayMusicController class by adding a new StreamController and StreamSink for loop status. It also adds a new method onTapLoop() that toggles the audio player's release mode between loop and release. The loop status is then added to the loopStatusInputData stream. Additionally, the CustomButtonControllerPlayMusicScreen widget is updated to include a new onTapLoop callback and a new IconButton that displays the loop icon in blue when the audio player is in loop mode. The inactive color of the slider is also updated to a blue color.